DQD SQL fully working on DataBricks
A python notebook is used to generate dbt models from the SQL queries generated using the DataQualityDashboard R library. See `scripts/notebooks/dbt_dqd_model_generator.ipynb` for more details on how this is achieved.

This is a horrible workaround and ideally the R library needs to be updated to work with DataBricks or ported to Python.